+++ src/add.js
@@ -1,0 +1,36 @@
+import ToDo from './todo'
+import Storage from './storage'
+import show from './display'
+const add = (() => {
+     function addProjectToProjects () {
+    let myProject = JSON.parse(localStorage.getItem('myProject'))
+  
+    const pTitle = document.getElementById('ptitle').value
+    const project = new Project(pTitle)
+    myProject.push(project)
+    Storage.storageMyProjects(myProject)
+  }
+
+  function addTask (eId,taskList, taskForm) {
+    let myProject = JSON.parse(localStorage.getItem('myProject'))
+    let tasks = myProject[eId]['_task']
+    const inTaskTitle = document.getElementById('titleInput').value
+    const inTaskDesc = document.getElementById('descriptionInput').value
+    const inTaskDate = document.getElementById('dateInput').value
+    const inTaskPriority = document.getElementById('priorityInput').value
+    const inTaskCheck = document.getElementById('checkInput').value
+  
+    const task = new ToDo (inTaskTitle, inTaskDesc, inTaskDate, inTaskPriority, inTaskCheck)
+  
+    tasks.push(task)
+    Storage.storageMyProjects(myProject)
+
+    return (myProject)
+  }
+  return {
+      addTask
+  }
+ 
+})()
+
+export default add
